+++ packages/enterprise/eslint.config.mjs
@@ -9,7 +9,11 @@ export default [
         'error',
         {
           ignoredFiles: ['{projectRoot}/eslint.config.{js,cjs,mjs,ts,cts,mts}'],
-          ignoredDependencies: ['@angular/core'],
+          ignoredDependencies: [
+            '@angular/core',
+            'tslib',
+            'jest-preset-angular',
+          ],
         },
       ],
     },